Recreate legacyrecovery type here for future use.
