+++ Cargo.toml
@@ -39,6 +39,7 @@ libra-tower = { path = "tools/tower" }
 libra-types = { path = "types" }
 libra-txs = { path = "tools/txs" }
 libra-wallet = { path = "tools/wallet" }
+vdf = { git = "https://github.com/0o-de-lally/verifiable_delay.git" }
 
 # Please do not add any test features here: they should be declared by the 
 # individual crate.
@@ -114,13 +115,6 @@ ureq = { version = "1.5.4", features = [
   "native-tls",
 ], default_features = false }
 url = { version = "2.3.1", features = ["serde"] }
-
-# Libra dependencies
-# ol-keys = { git = "https://github.com/simsekgokhan/libra.git", branch = "v6-dwh-err" }
-ol-types = { git = "https://github.com/simsekgokhan/libra.git", branch = "v6-dwh-err" }
-# diem-wallet = { git = "https://github.com/simsekgokhan/libra.git", branch = "v6-dwh-err" }
-# diem-types = { git = "https://github.com/simsekgokhan/libra.git", branch = "v6-dwh-err" }
-vdf = { git = "https://github.com/0o-de-lally/verifiable_delay.git" }
 
 
 ### MOVE DEPENDENCIES
