+++ tools/genesis/Cargo.toml
@@ -18,12 +18,9 @@ libra-framework = { workspace = true }
 libra-types = { workspace = true }
 libra-wallet = { workspace = true }
 
-
 clap = { workspace = true }
 anyhow = { workspace = true }
-ol-types = { workspace = true }
 reqwest = { workspace = true }
-# diem-types = { workspace = true }
 base64 = { workspace = true }
 bcs = { workspace = true }
 hex = { workspace = true }
