+++ tools/genesis/src/compare.rs
@@ -7,8 +7,8 @@ use crate::db_utils;
 use anyhow;
 use diem_types::account_address::AccountAddress;
 use diem_types::account_state::AccountState;
-use ol_types::legacy_recovery::LegacyRecovery;
-use ol_types::legacy_recovery::read_from_recovery_file;
+use libra_types::legacy_recovery::LegacyRecovery;
+use libra_types::legacy_recovery::read_from_recovery_file;
 use std::convert::TryFrom;
 use std::path::PathBuf;
 use ol_types::OLProgress;
