+++ types/src/legacy_types/mod.rs
@@ -2,6 +2,7 @@
 
 // pub mod account;
 // pub mod autopay;
+pub mod ancestry;
 pub mod app_cfg;
 pub mod block;
 pub mod mode_ol;
@@ -10,6 +11,9 @@ pub mod tx_error;
 pub mod vdf_difficulty;
 pub mod network_playlist;
 pub mod makewhole_resource;
+pub mod legacy_address;
+pub mod legacy_recovery;
+pub mod legacy_currency_info;
 // pub mod config;
 // pub mod dialogue;
 // pub mod epoch_timer;
@@ -20,12 +24,13 @@ pub mod makewhole_resource;
 
 // pub mod pay_instruction;
 // pub mod rpc_playlist;
-// pub mod validator_config;
-// pub mod wallet;
+pub mod legacy_miner_state;
+pub mod validator_config;
+pub mod wallet;
 // pub mod legacy_recovery;
 // pub mod ancestry;
-// pub mod receipts;
-// pub mod cumulative_deposits;
+pub mod receipts;
+pub mod cumulative_deposits;
 // pub mod cli_progress;
 // pub use cli_progress::OLProgress;
