+++ types/src/legacy_types/receipts.rs
@@ -1,12 +1,10 @@
 //! fullnode counter for system address
 
 use anyhow::Result;
-use diem_types::{access_path::AccessPath, account_config::constants::CORE_CODE_ADDRESS};
 use move_core_types::account_address::AccountAddress;
 use move_core_types::{
     ident_str,
     identifier::IdentStr,
-    language_storage::{ResourceKey, StructTag},
     move_resource::MoveStructType,
 };
 use serde::{Deserialize, Serialize};
@@ -30,24 +28,6 @@ impl MoveStructType for ReceiptsResource {
 }
 
 impl ReceiptsResource {
-    ///
-    pub fn struct_tag() -> StructTag {
-        StructTag {
-            address: CORE_CODE_ADDRESS,
-            module: ReceiptsResource::module_identifier(),
-            name: ReceiptsResource::struct_identifier(),
-            type_params: vec![],
-        }
-    }
-    ///
-    pub fn access_path(account: AccountAddress) -> AccessPath {
-        let resource_key = ResourceKey::new(account, ReceiptsResource::struct_tag());
-        AccessPath::resource_access_path(resource_key)
-    }
-    ///
-    pub fn resource_path() -> Vec<u8> {
-        AccessPath::resource_access_vec(ReceiptsResource::struct_tag())
-    }
 
     ///
     pub fn try_from_bytes(bytes: &[u8]) -> Result<Self> {
