+++ types/src/legacy_types/validator_config.rs
@@ -1,13 +1,11 @@
 //! validator config view for web monitor
 
 use anyhow::Result;
-use diem_types::{access_path::AccessPath, account_config::constants::CORE_CODE_ADDRESS};
 use move_core_types::account_address::AccountAddress;
 use move_core_types::{
     ident_str,
     identifier::IdentStr,
-    language_storage::{ResourceKey, StructTag},
-    move_resource::{MoveResource, MoveStructType},
+    move_resource::MoveStructType,
 };
 use serde::{Deserialize, Serialize};
 
@@ -40,40 +38,23 @@ impl MoveStructType for ValidatorConfigResource {
     const MODULE_NAME: &'static IdentStr = ident_str!("ValidatorConfig");
     const STRUCT_NAME: &'static IdentStr = ident_str!("ValidatorConfig");
 }
-impl MoveResource for ValidatorConfigResource {}
+// impl MoveResource for ValidatorConfigResource {}
 
 impl ValidatorConfigResource {
-    ///
-    pub fn struct_tag() -> StructTag {
-        StructTag {
-            address: CORE_CODE_ADDRESS,
-            module: ValidatorConfigResource::module_identifier(),
-            name: ValidatorConfigResource::struct_identifier(),
-            type_params: vec![],
-        }
-    }
-    ///
-    pub fn access_path(account: AccountAddress) -> AccessPath {
-        let resource_key = ResourceKey::new(account, ValidatorConfigResource::struct_tag());
-        AccessPath::resource_access_path(resource_key)
-    }
-    ///
-    pub fn resource_path() -> Vec<u8> {
-        AccessPath::resource_access_vec(ValidatorConfigResource::struct_tag())
-    }
+
 
     ///
     pub fn try_from_bytes(bytes: &[u8]) -> Result<Self> {
         bcs::from_bytes(bytes).map_err(Into::into)
     }
 
-    ///
-    pub fn get_view(&self) -> ValidatorConfigView {
-        ValidatorConfigView {
-            operator_account: self.operator_account.clone(),
-            operator_has_balance: None,
-        }
-    }
+    // ///
+    // pub fn get_view(&self) -> ValidatorConfigView {
+    //     ValidatorConfigView {
+    //         operator_account: self.operator_account.clone(),
+    //         operator_has_balance: None,
+    //     }
+    // }
 }
 
 /// Struct that represents a view for Validator Config view
