+++ types/src/legacy_types/wallet.rs
@@ -1,12 +1,11 @@
 //! community wallet resource
 
 use anyhow::Result;
-use diem_types::{access_path::AccessPath, account_config::constants::CORE_CODE_ADDRESS};
+// use diem_types::{access_path::AccessPath, account_config::constants::CORE_CODE_ADDRESS};
 use move_core_types::account_address::AccountAddress;
 use move_core_types::{
     ident_str,
     identifier::IdentStr,
-    language_storage::{ResourceKey, StructTag},
     move_resource::MoveStructType,
 };
 use serde::{Deserialize, Serialize};
@@ -26,24 +25,24 @@ impl MoveStructType for CommunityWalletsResourceLegacy {
 }
 
 impl CommunityWalletsResourceLegacy {
+    // ///
+    // pub fn struct_tag() -> StructTag {
+    //     StructTag {
+    //         address: CORE_CODE_ADDRESS,
+    //         module: CommunityWalletsResourceLegacy::module_identifier(),
+    //         name: CommunityWalletsResourceLegacy::struct_identifier(),
+    //         type_params: vec![],
+    //     }
+    // }
+    // ///
+    // pub fn access_path(account: AccountAddress) -> AccessPath {
+    //     let resource_key = ResourceKey::new(account, CommunityWalletsResourceLegacy::struct_tag());
+    //     AccessPath::resource_access_path(resource_key)
+    // }
     ///
-    pub fn struct_tag() -> StructTag {
-        StructTag {
-            address: CORE_CODE_ADDRESS,
-            module: CommunityWalletsResourceLegacy::module_identifier(),
-            name: CommunityWalletsResourceLegacy::struct_identifier(),
-            type_params: vec![],
-        }
-    }
-    ///
-    pub fn access_path(account: AccountAddress) -> AccessPath {
-        let resource_key = ResourceKey::new(account, CommunityWalletsResourceLegacy::struct_tag());
-        AccessPath::resource_access_path(resource_key)
-    }
-    ///
-    pub fn resource_path() -> Vec<u8> {
-        AccessPath::resource_access_vec(CommunityWalletsResourceLegacy::struct_tag())
-    }
+    // pub fn resource_path() -> Vec<u8> {
+    //     AccessPath::resource_access_vec(CommunityWalletsResourceLegacy::struct_tag())
+    // }
 
     ///
     pub fn try_from_bytes(bytes: &[u8]) -> Result<Self> {
@@ -66,24 +65,24 @@ impl MoveStructType for SlowWalletResource {
 }
 
 impl SlowWalletResource {
-    ///
-    pub fn struct_tag() -> StructTag {
-        StructTag {
-            address: CORE_CODE_ADDRESS,
-            module: SlowWalletResource::module_identifier(),
-            name: SlowWalletResource::struct_identifier(),
-            type_params: vec![],
-        }
-    }
-    ///
-    pub fn access_path(account: AccountAddress) -> AccessPath {
-        let resource_key = ResourceKey::new(account, SlowWalletResource::struct_tag());
-        AccessPath::resource_access_path(resource_key)
-    }
-    ///
-    pub fn resource_path() -> Vec<u8> {
-        AccessPath::resource_access_vec(SlowWalletResource::struct_tag())
-    }
+    //
+    // pub fn struct_tag() -> StructTag {
+    //     StructTag {
+    //         address: CORE_CODE_ADDRESS,
+    //         module: SlowWalletResource::module_identifier(),
+    //         name: SlowWalletResource::struct_identifier(),
+    //         type_params: vec![],
+    //     }
+    // }
+    // ///
+    // pub fn access_path(account: AccountAddress) -> AccessPath {
+    //     let resource_key = ResourceKey::new(account, SlowWalletResource::struct_tag());
+    //     AccessPath::resource_access_path(resource_key)
+    // }
+    // ///
+    // pub fn resource_path() -> Vec<u8> {
+    //     AccessPath::resource_access_vec(SlowWalletResource::struct_tag())
+    // }
 
     ///
     pub fn try_from_bytes(bytes: &[u8]) -> Result<Self> {
@@ -105,24 +104,24 @@ impl MoveStructType for SlowWalletListResource {
 }
 
 impl SlowWalletListResource {
-    ///
-    pub fn struct_tag() -> StructTag {
-        StructTag {
-            address: CORE_CODE_ADDRESS,
-            module: SlowWalletListResource::module_identifier(),
-            name: SlowWalletListResource::struct_identifier(),
-            type_params: vec![],
-        }
-    }
-    ///
-    pub fn access_path(account: AccountAddress) -> AccessPath {
-        let resource_key = ResourceKey::new(account, SlowWalletListResource::struct_tag());
-        AccessPath::resource_access_path(resource_key)
-    }
-    ///
-    pub fn resource_path() -> Vec<u8> {
-        AccessPath::resource_access_vec(SlowWalletListResource::struct_tag())
-    }
+    // ///
+    // pub fn struct_tag() -> StructTag {
+    //     StructTag {
+    //         address: CORE_CODE_ADDRESS,
+    //         module: SlowWalletListResource::module_identifier(),
+    //         name: SlowWalletListResource::struct_identifier(),
+    //         type_params: vec![],
+    //     }
+    // }
+    // ///
+    // pub fn access_path(account: AccountAddress) -> AccessPath {
+    //     let resource_key = ResourceKey::new(account, SlowWalletListResource::struct_tag());
+    //     AccessPath::resource_access_path(resource_key)
+    // }
+    // ///
+    // pub fn resource_path() -> Vec<u8> {
+    //     AccessPath::resource_access_vec(SlowWalletListResource::struct_tag())
+    // }
 
     ///
     pub fn try_from_bytes(bytes: &[u8]) -> Result<Self> {
